+++ src/Component/Footer.tsx
@@ -1,4 +1,9 @@
 import styled from "styled-components";
+function Footer() {
+  return <Footerstyle>만든이 : 차차</Footerstyle>;
+}
+
+export default Footer;
 
 const Footerstyle = styled.div`
   margin: 0;
@@ -9,5 +14,3 @@ const Footerstyle = styled.div`
   align-items: center;
   margin: 30px 0px 30px 0px;
 `;
-
-export { Footerstyle };
