--- src/Component/Footer/Footer.tsx
@@ -1,7 +1,0 @@
-import { Footerstyle } from "./style";
-
-function Footer() {
-  return <Footerstyle>만든이 : 차차</Footerstyle>;
-}
-
-export default Footer;
