--- src/Component/Profill/style.js
@@ -1,13 +1,0 @@
-import styled from "styled-components";
-
-const ProfileImg = styled.img`
-  position: relative;
-  margin: auto;
-  width: 150px;
-  height: 150px;
-  border: 1px solid #d3d3d3;
-  border-radius: 50%;
-  cursor: pointer;
-`;
-
-export { ProfileImg };
