+++ src/Router.tsx
@@ -1,19 +1,19 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Header from "./Component/Header/Header";
-import Loginpage from "./Page/Login/Login";
-import HomePage from "./Page/Home/Home.Page";
-import Mypage from "./Page/MyPage/MyPage";
-import SignupPage from "./Page/Signup/Signup";
+import Header from "./Component/Header";
+import LoginPage from "./Page/LoginPage";
+import MainPage from "./Page/MainPage";
+import Mypage from "./Page/MyPage";
+import SignupPage from "./Page/Signup";
 import styled from "styled-components";
-import Footer from "./Component/Footer/Footer";
+import Footer from "./Component/Footer";
 function Router() {
   return (
     <BrowserRouter>
       <Layout>
         <Header />
         <Routes>
-          <Route path="/" element={<HomePage />} />
-          <Route path="Login" element={<Loginpage />} />
+          <Route path="/" element={<MainPage />} />
+          <Route path="Login" element={<LoginPage />} />
           <Route path="Signup" element={<SignupPage />} />
           <Route path="MyPage" element={<Mypage />} />
         </Routes>
