--- src/page/Home/Home.Page.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import Footer from "../../Component/Footer/Footer";
-import Main from "../../Component/Main/Main";
-import { HomePageLayout } from "./style";
-
-function HomePage() {
-  return (
-    <HomePageLayout>
-      <Main />
-    </HomePageLayout>
-  );
-}
-
-export default HomePage;
