--- src/page/Home/style.js
@@ -1,13 +1,0 @@
-import styled from "styled-components";
-
-const HomePageLayout = styled.div`
-  width: 100%;
-  height: 90%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: #f3f3f3;
-`;
-
-export { HomePageLayout };
